# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Swagger documentation for the Gin-Golang project.",
        "title": "Gin-Golang API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Đăng nhập tài khoản người dùng mới với email, mật khẩu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập",
                "parameters": [
                    {
                        "description": "Thông tin người dùng",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginInDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.TokenResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Tạo tài khoản người dùng mới với email, mật khẩu và tên",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký người dùng mới",
                "parameters": [
                    {
                        "description": "Thông tin người dùng",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Người dùng được tạo thành công",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
        "/users/avatar": {
            "put": {
                "description": "Upload file avatar (image) cho user - max 10MB",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Upload avatar **FILE**",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của user",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File avatar để upload",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy thông tin user đang đăng nhập",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Trả thông tin user đang đăng nhập  **Yêu cầu xác thực**",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.DataResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "dto.LoginInDto": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
...
        "dto.RegisterDto": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "dto.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}