# internal/dto/user.go
package dto

type RegisterDto struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=3"`
	Name     string `json:"name" binding:"required,min=3"`
}

type LoginInDto struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=3"`
}